+++ certbot-dns-gehirn/certbot_dns_gehirn/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_gehirn.dns_gehirn` plugin.
