+++ certbot-dns-gehirn/docs/api/dns_gehirn.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_gehirn.dns_gehirn`
+:mod:`certbot_dns_gehirn._internal.dns_gehirn`
 ------------------------------------
 
-.. automodule:: certbot_dns_gehirn.dns_gehirn
+.. automodule:: certbot_dns_gehirn._internal.dns_gehirn
    :members:
